:bug: use absolute path in singularity install

+++ .travis/install_singularity.sh
@@ -1,12 +1,12 @@
 #!/bin/bash -ex
 
-# if [ -f "$HOME/singularity/bin/singularity" ]; then
-#     echo exists
-# else
-apt-get update && apt-get install squashfs-tools dh-autoreconf build-essential
-git clone https://github.com/singularityware/singularity.git
-cd singularity
-./autogen.sh
-./configure --prefix=$HOME/singularity
-make install
-# fi
+if [ -f "/home/travis/build/leukgen/toil_container/singularity/bin/singularity" ]; then
+    echo exists
+else
+    apt-get update && apt-get install squashfs-tools dh-autoreconf build-essential
+    git clone https://github.com/singularityware/singularity.git
+    cd singularity
+    ./autogen.sh
+    ./configure --prefix=/home/travis/build/leukgen/toil_container/singularity
+    make install
+fi
